Replace training_mod_tui with training_mod_tui_2, update byteflags dependency for random selection
