+++ training_mod_tui/Cargo.toml
@@ -1,18 +1,18 @@
 [package]
 name = "training_mod_tui"
 version = "0.1.0"
-edition = "2018"
+edition = "2021"
 
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+[lib]
+doctest = false
 
 [dependencies]
-tui = { version = "0.16.0", default-features = false }
-unicode-width = "0.1.9"
-training_mod_consts = { path = "../training_mod_consts", default-features = false}
-serde_json = "1.0.79"
-bitflags = "1.2.1"
+itertools = "0.11.0"
+ratatui = { git = "https://github.com/tonogdlp/ratatui.git", branch = "single-cell", default-features = false}
+serde = { version = "1.0.188", features = ["derive"] }
+serde_json = "1.0.106"
 crossterm = { version = "0.22.1", optional = true }
 
 [features]
 default = []
-has_terminal = ["crossterm", "tui/crossterm"]
+has_terminal = ["crossterm", "ratatui/crossterm"]
