+++ training_mod_tui/src/containers/app.rs
