+++ training_mod_tui/src/containers/toggle.rs
