+++ training_mod_tui/src/main.rs
@@ -1,334 +1,206 @@
-#[cfg(feature = "has_terminal")]
 use crossterm::{
-    event::{self, DisableMouseCapture, EnableMouseCapture, Event, KeyCode},
+    event::{self, EnableMouseCapture, Event, KeyCode},
     execute,
     terminal::{disable_raw_mode, enable_raw_mode, EnterAlternateScreen, LeaveAlternateScreen},
 };
-
+use ratatui::backend::CrosstermBackend;
+use ratatui::Terminal;
 use std::error::Error;
-#[cfg(feature = "has_terminal")]
 use std::{
     io,
     time::{Duration, Instant},
 };
-#[cfg(feature = "has_terminal")]
-use tui::backend::CrosstermBackend;
-use tui::Terminal;
 
-use training_mod_consts::*;
-
-fn test_backend_setup(
-    ui_menu: UiMenu,
-    menu_defaults: (UiMenu, String),
-) -> Result<
-    (
-        Terminal<training_mod_tui::TestBackend>,
-        training_mod_tui::App,
-    ),
-    Box<dyn Error>,
-> {
-    let app = training_mod_tui::App::new(ui_menu, menu_defaults);
-    let backend = tui::backend::TestBackend::new(120, 15);
-    let terminal = Terminal::new(backend)?;
-    let mut state = tui::widgets::ListState::default();
-    state.select(Some(1));
-
-    Ok((terminal, app))
-}
-
-#[test]
-fn test_set_airdodge() -> Result<(), Box<dyn Error>> {
-    let menu;
-    let mut prev_menu;
-    let menu_defaults;
-    unsafe {
-        prev_menu = MENU.clone();
-        menu = ui_menu(MENU);
-        menu_defaults = (ui_menu(MENU), serde_json::to_string(&MENU).unwrap());
-    }
-
-    let (_terminal, mut app) = test_backend_setup(menu, menu_defaults)?;
-    // Enter Mash Section
-    app.next_tab();
-    // Enter Mash Toggles
-    app.on_a();
-    // Set Mash Airdodge
-    app.on_a();
-    let menu_json = app.get_menu_selections();
-    let menu_struct = serde_json::from_str::<MenuJsonStruct>(&menu_json).unwrap();
-    let menu = menu_struct.menu;
-    let _ = menu_struct.defaults_menu;
-    prev_menu.mash_state.toggle(Action::AIR_DODGE);
-    assert_eq!(
-        serde_json::to_string(&prev_menu).unwrap(),
-        serde_json::to_string(&menu).unwrap()
-    );
-
-    Ok(())
-}
-
-#[test]
-fn test_ensure_menu_retains_selections() -> Result<(), Box<dyn Error>> {
-    let menu;
-    let prev_menu;
-    let menu_defaults;
-    unsafe {
-        prev_menu = MENU.clone();
-        menu = ui_menu(MENU);
-        menu_defaults = (ui_menu(MENU), serde_json::to_string(&MENU).unwrap());
-    }
-
-    let (_terminal, app) = test_backend_setup(menu, menu_defaults)?;
-    let menu_json = app.get_menu_selections();
-    let menu_struct = serde_json::from_str::<MenuJsonStruct>(&menu_json).unwrap();
-    let menu = menu_struct.menu;
-    let _ = menu_struct.defaults_menu;
-    // At this point, we didn't change the menu at all; we should still see all the same options.
-    assert_eq!(
-        serde_json::to_string(&prev_menu).unwrap(),
-        serde_json::to_string(&menu).unwrap()
-    );
-
-    Ok(())
-}
-
-#[test]
-fn test_save_and_reset_defaults() -> Result<(), Box<dyn Error>> {
-    let menu;
-    let mut prev_menu;
-    let menu_defaults;
-    unsafe {
-        prev_menu = MENU.clone();
-        menu = ui_menu(MENU);
-        menu_defaults = (ui_menu(MENU), serde_json::to_string(&MENU).unwrap());
-    }
-
-    let (_terminal, mut app) = test_backend_setup(menu, menu_defaults)?;
-
-    // Enter Mash Section
-    app.next_tab();
-    // Enter Mash Toggles
-    app.on_a();
-    // Set Mash Airdodge
-    app.on_a();
-    // Return to submenu selection
-    app.on_b();
-    // Save Defaults
-    app.save_defaults();
-    // Enter Mash Toggles again
-    app.on_a();
-    // Unset Mash Airdodge
-    app.on_a();
-
-    let menu_json = app.get_menu_selections();
-    let menu_struct = serde_json::from_str::<MenuJsonStruct>(&menu_json).unwrap();
-    let menu = menu_struct.menu;
-    let defaults_menu = menu_struct.defaults_menu;
-
-    assert_eq!(
-        serde_json::to_string(&prev_menu).unwrap(),
-        serde_json::to_string(&menu).unwrap(),
-        "The menu should be the same as how we started"
-    );
-    prev_menu.mash_state.toggle(Action::AIR_DODGE);
-    assert_eq!(
-        serde_json::to_string(&prev_menu).unwrap(),
-        serde_json::to_string(&defaults_menu).unwrap(),
-        "The defaults menu should have Mash Airdodge"
-    );
-
-    // Reset current menu alone to defaults
-    app.reset_current_submenu();
-    let menu_json = app.get_menu_selections();
-    let menu_struct = serde_json::from_str::<MenuJsonStruct>(&menu_json).unwrap();
-    let menu = menu_struct.menu;
-    let _ = menu_struct.defaults_menu;
-
-    assert_eq!(
-        serde_json::to_string(&prev_menu).unwrap(),
-        serde_json::to_string(&menu).unwrap(),
-        "The menu should have Mash Airdodge"
-    );
-
-    // Enter Mash Section
-    app.next_tab();
-    // Enter Mash Toggles
-    app.on_a();
-    // Unset Mash Airdodge
-    app.on_a();
-    // Return to submenu selection
-    app.on_b();
-    // Go down to Followup Toggles
-    app.on_down();
-    // Enter Followup Toggles
-    app.on_a();
-    // Go down and set Jump
-    app.on_down();
-    app.on_a();
-    // Return to submenu selection
-    app.on_b();
-    // Save defaults
-    app.save_defaults();
-    // Go back in and unset Jump
-    app.on_a();
-    app.on_down();
-    app.on_a();
-    // Return to submenu selection
-    app.on_b();
-    // Reset all to defaults
-    app.reset_all_submenus();
-    let menu_json = app.get_menu_selections();
-    let menu_struct = serde_json::from_str::<MenuJsonStruct>(&menu_json).unwrap();
-    let menu = menu_struct.menu;
-    let _ = menu_struct.defaults_menu;
-
-    prev_menu.mash_state.toggle(Action::AIR_DODGE);
-    prev_menu.follow_up.toggle(Action::JUMP);
-    assert_eq!(
-        serde_json::to_string(&prev_menu).unwrap(),
-        serde_json::to_string(&menu).unwrap(),
-        "The menu should have Mash Airdodge off and Followup Jump on"
-    );
-
-    Ok(())
-}
-
-fn _get_frame_buffer(
-    mut terminal: Terminal<training_mod_tui::TestBackend>,
-    mut app: training_mod_tui::App,
-) -> Result<String, Box<dyn Error>> {
-    let frame_res = terminal.draw(|f| training_mod_tui::ui(f, &mut app))?;
-    let mut full_frame_buffer = String::new();
-    for (i, cell) in frame_res.buffer.content().iter().enumerate() {
-        full_frame_buffer.push_str(&cell.symbol);
-        if i % frame_res.area.width as usize == frame_res.area.width as usize - 1 {
-            full_frame_buffer.push_str("\n");
-        }
-    }
-    full_frame_buffer.push_str("\n");
-
-    Ok(full_frame_buffer)
-}
-
-#[test]
-fn test_toggle_naming() -> Result<(), Box<dyn Error>> {
-    let menu;
-    let mut prev_menu;
-    let menu_defaults;
-    unsafe {
-        prev_menu = MENU.clone();
-        menu = ui_menu(MENU);
-        menu_defaults = (ui_menu(MENU), serde_json::to_string(&MENU).unwrap());
-    }
-
-    let (mut terminal, mut app) = test_backend_setup(menu, menu_defaults)?;
-    // Enter Mash Section
-    app.next_tab();
-    // Enter Mash Toggles
-    app.on_a();
-    // Set Mash Airdodge
-    app.on_a();
-
-    let frame_buffer = _get_frame_buffer(terminal, app)?;
-    assert!(frame_buffer.contains("Airdodge"));
-
-    Ok(())
-}
+use training_mod_tui::{
+    App, AppPage, InputControl, StatefulList, StatefulSlider, StatefulTable, SubMenu, SubMenuType,
+    Tab, Toggle, NX_SUBMENU_COLUMNS, NX_SUBMENU_ROWS,
+};
 
 fn main() -> Result<(), Box<dyn Error>> {
-    let args: Vec<String> = std::env::args().collect();
-    let inputs = args.get(1);
-    let menu;
-    let menu_defaults;
+    let mut app = create_app();
+    let json = "{\"Menu Open Start Press\":[1,0],\"Dmg Range (CPU)\":[40,100]}";
+    app.update_from_json(json);
+    let mut terminal = setup_terminal()?;
 
-    unsafe {
-        menu = ui_menu(MENU);
-        menu_defaults = (ui_menu(MENU), serde_json::to_string(&MENU).unwrap());
-    }
+    let tick_rate = Duration::from_millis(250);
+    let res = run_app(&mut terminal, app, tick_rate);
+    restore_terminal(terminal)?;
 
-    #[cfg(not(feature = "has_terminal"))]
-    {
-        let (mut terminal, mut app) = test_backend_setup(menu, menu_defaults)?;
-        if inputs.is_some() {
-            inputs
-                .unwrap()
-                .split(",")
-                .for_each(|input| match input.to_uppercase().as_str() {
-                    "X" => app.save_defaults(),
-                    "Y" => app.reset_current_submenu(),
-                    "Z" => app.reset_all_submenus(),
-                    "L" => app.previous_tab(),
-                    "R" => app.next_tab(),
-                    "A" => app.on_a(),
-                    "B" => app.on_b(),
-                    "UP" => app.on_up(),
-                    "DOWN" => app.on_down(),
-                    "LEFT" => app.on_left(),
-                    "RIGHT" => app.on_right(),
-                    _ => {}
-                })
-        }
-        let frame_res = terminal.draw(|f| training_mod_tui::ui(f, &mut app))?;
-        let menu_json = app.get_menu_selections();
-
-        for (i, cell) in frame_res.buffer.content().iter().enumerate() {
-            print!("{}", cell.symbol);
-            if i % frame_res.area.width as usize == frame_res.area.width as usize - 1 {
-                println!();
-            }
-        }
-        println!();
-
-        println!("Menu:\n{menu_json}");
-    }
-
-    #[cfg(feature = "has_terminal")]
-    {
-        let app = training_mod_tui::App::new(menu, menu_defaults);
-
-        // setup terminal
-        enable_raw_mode()?;
-        let mut stdout = io::stdout();
-        execute!(stdout, EnterAlternateScreen, EnableMouseCapture)?;
-        let backend = CrosstermBackend::new(stdout);
-        let mut terminal = Terminal::new(backend)?;
-
-        let tick_rate = Duration::from_millis(250);
-        let res = run_app(&mut terminal, app, tick_rate);
-
-        // restore terminal
-        disable_raw_mode()?;
-        execute!(
-            terminal.backend_mut(),
-            LeaveAlternateScreen,
-            DisableMouseCapture
-        )?;
-        terminal.show_cursor()?;
-
-        if let Err(err) = res {
-            println!("{:?}", err)
-        } else {
-            println!("JSONs: {:#?}", res.as_ref().unwrap());
-            unsafe {
-                let menu = serde_json::from_str::<MenuJsonStruct>(&res.as_ref().unwrap()).unwrap();
-                println!("menu: {:#?}", menu);
-            }
-        }
+    if let Err(err) = res {
+        println!("Error: {:?}", err)
+    } else {
+        println!("JSON: {:#?}", res.as_ref().unwrap());
     }
 
     Ok(())
 }
 
-#[cfg(feature = "has_terminal")]
-fn run_app<B: tui::backend::Backend>(
+fn setup_terminal() -> Result<Terminal<CrosstermBackend<io::Stdout>>, Box<dyn Error>> {
+    enable_raw_mode()?;
+    let mut stdout = io::stdout();
+    execute!(stdout, EnterAlternateScreen, EnableMouseCapture)?;
+    let backend = CrosstermBackend::new(stdout);
+    let terminal = Terminal::new(backend)?;
+    Ok(terminal)
+}
+
+fn restore_terminal(
+    mut terminal: Terminal<CrosstermBackend<io::Stdout>>,
+) -> Result<(), Box<dyn Error>> {
+    disable_raw_mode()?;
+    execute!(terminal.backend_mut(), LeaveAlternateScreen)?;
+    terminal.show_cursor()?;
+    Ok(())
+}
+
+pub fn create_app<'a>() -> App<'a> {
+    let mut app = App::new();
+    let mut button_tab_submenus: Vec<SubMenu> = Vec::new();
+    button_tab_submenus.push(SubMenu {
+        title: "Menu Open Start Press",
+        id: "menu_open_start_press",
+        help_text: "Menu Open Start Press: Should pressing start open the menu?",
+        submenu_type: SubMenuType::ToggleSingle,
+        toggles: new_toggle_table(new_on_off()),
+        slider: None,
+    });
+    button_tab_submenus.push(SubMenu {
+        title: "Save State Save",
+        id: "save_state_save",
+        help_text: "Save State Save: Hold any one button and press the others to trigger",
+        submenu_type: SubMenuType::ToggleMultiple,
+        toggles: new_toggle_table(new_button_combo()),
+        slider: None,
+    });
+    button_tab_submenus.push(SubMenu {
+        title: "Save State Load",
+        id: "save_state_load",
+        help_text: "Save State Load: Hold any one button and press the others to trigger",
+        submenu_type: SubMenuType::ToggleMultiple,
+        toggles: new_toggle_table(new_button_combo()),
+        slider: None,
+    });
+    button_tab_submenus.push(SubMenu {
+        title: "Input Record",
+        id: "input_record",
+        help_text: "Input Record: Hold any one button and press the others to trigger",
+        submenu_type: SubMenuType::ToggleMultiple,
+        toggles: new_toggle_table(new_button_combo()),
+        slider: None,
+    });
+    button_tab_submenus.push(SubMenu {
+        title: "Input Playback",
+        id: "input_playback",
+        help_text: "Input Playback: Hold any one button and press the others to trigger",
+        submenu_type: SubMenuType::ToggleMultiple,
+        toggles: new_toggle_table(new_button_combo()),
+        slider: None,
+    });
+
+    let button_tab = Tab {
+        id: "button",
+        title: "Button Config",
+        submenus: StatefulTable::with_items(
+            NX_SUBMENU_ROWS,
+            NX_SUBMENU_COLUMNS,
+            button_tab_submenus.clone(),
+        ),
+    };
+    let button_tab_2 = Tab {
+        id: "button",
+        title: "Button Config 2",
+        submenus: StatefulTable::with_items(
+            NX_SUBMENU_ROWS,
+            NX_SUBMENU_COLUMNS,
+            button_tab_submenus.clone(),
+        ),
+    };
+    let button_tab_3 = Tab {
+        id: "button",
+        title: "Button Config 3",
+        submenus: StatefulTable::with_items(
+            NX_SUBMENU_ROWS,
+            NX_SUBMENU_COLUMNS,
+            button_tab_submenus.clone(),
+        ),
+    };
+    let button_tab_4 = Tab {
+        id: "button",
+        title: "Button Config 4",
+        submenus: StatefulTable::with_items(
+            NX_SUBMENU_ROWS,
+            NX_SUBMENU_COLUMNS,
+            button_tab_submenus.clone(),
+        ),
+    };
+    let button_tab_5 = Tab {
+        id: "button",
+        title: "Button Config 5",
+        submenus: StatefulTable::with_items(
+            NX_SUBMENU_ROWS,
+            NX_SUBMENU_COLUMNS,
+            button_tab_submenus.clone(),
+        ),
+    };
+
+    let mut save_state_tab_submenus: Vec<SubMenu> = Vec::new();
+    save_state_tab_submenus.push(SubMenu {
+        title: "Mirroring",
+        id: "save_state_mirroring",
+        help_text:
+            "Mirroring: Flips save states in the left-right direction across the stage center",
+        submenu_type: SubMenuType::ToggleSingle,
+        toggles: new_toggle_table(new_on_off()),
+        slider: None,
+    });
+    save_state_tab_submenus.push(SubMenu {
+        title: "Auto Save States",
+        id: "save_state_autoload",
+        help_text: "Auto Save States: Load save state when any fighter dies",
+        submenu_type: SubMenuType::ToggleSingle,
+        toggles: new_toggle_table(new_on_off()),
+        slider: None,
+    });
+    save_state_tab_submenus.push(SubMenu {
+        title: "Dmg Range (CPU)",
+        id: "save_damage_limits_cpu",
+        help_text: "Limits on random damage to apply to the CPU when loading a save state",
+        submenu_type: SubMenuType::Slider,
+        toggles: StatefulTable::new(NX_SUBMENU_ROWS, NX_SUBMENU_COLUMNS),
+        slider: Some(StatefulSlider::new()),
+    });
+    let save_states_tab = Tab {
+        id: "save_state",
+        title: "Save States",
+        submenus: StatefulTable::with_items(
+            NX_SUBMENU_ROWS,
+            NX_SUBMENU_COLUMNS,
+            save_state_tab_submenus,
+        ),
+    };
+
+    app.tabs = StatefulList::with_items(vec![
+        button_tab,
+        button_tab_2,
+        button_tab_3,
+        button_tab_4,
+        button_tab_5,
+        save_states_tab,
+    ]);
+    app
+}
+
+fn run_app<B: ratatui::backend::Backend>(
     terminal: &mut Terminal<B>,
     mut app: training_mod_tui::App,
     tick_rate: Duration,
 ) -> io::Result<String> {
     let mut last_tick = Instant::now();
     loop {
-        terminal.draw(|f| training_mod_tui::ui(f, &mut app).clone())?;
-        let menu_json = app.get_menu_selections();
+        terminal.draw(|f| training_mod_tui::render_ui(f, &mut app))?;
+        if app.page == AppPage::CLOSE {
+            return Ok(app.to_json());
+        }
 
         let timeout = tick_rate
             .checked_sub(last_tick.elapsed())
@@ -337,12 +209,15 @@ fn run_app<B: tui::backend::Backend>(
         if crossterm::event::poll(timeout)? {
             if let Event::Key(key) = event::read()? {
                 match key.code {
-                    KeyCode::Char('q') => return Ok(menu_json),
-                    KeyCode::Char('x') => app.save_defaults(),
-                    KeyCode::Char('p') => app.reset_current_submenu(),
-                    KeyCode::Char('o') => app.reset_all_submenus(),
-                    KeyCode::Char('r') => app.next_tab(),
-                    KeyCode::Char('l') => app.previous_tab(),
+                    KeyCode::Char('q') => app.page = AppPage::CLOSE,
+                    KeyCode::Char('a') => app.on_a(),
+                    KeyCode::Char('b') => app.on_b(),
+                    KeyCode::Char('x') => app.on_x(),
+                    KeyCode::Char('y') => app.on_y(),
+                    KeyCode::Char('o') => app.on_zl(),
+                    KeyCode::Char('p') => app.on_zr(),
+                    KeyCode::Char('l') => app.on_l(),
+                    KeyCode::Char('r') => app.on_r(),
                     KeyCode::Left => app.on_left(),
                     KeyCode::Right => app.on_right(),
                     KeyCode::Down => app.on_down(),
@@ -357,4 +232,99 @@ fn run_app<B: tui::backend::Backend>(
             last_tick = Instant::now();
         }
     }
+}
+
+fn new_button_combo<'a>() -> Vec<Toggle<'a>> {
+    let a_button = Toggle {
+        title: "A Button",
+        value: 0,
+        max: 1,
+    };
+    let b_button = Toggle {
+        title: "B Button",
+        value: 0,
+        max: 1,
+    };
+    let x_button = Toggle {
+        title: "X Button",
+        value: 0,
+        max: 1,
+    };
+    let y_button = Toggle {
+        title: "Y Button",
+        value: 0,
+        max: 1,
+    };
+    let l_button = Toggle {
+        title: "L Button",
+        value: 0,
+        max: 1,
+    };
+    let r_button = Toggle {
+        title: "R Button",
+        value: 0,
+        max: 1,
+    };
+    let zl_button = Toggle {
+        title: "ZL Button",
+        value: 0,
+        max: 1,
+    };
+    let zr_button = Toggle {
+        title: "ZR Button",
+        value: 0,
+        max: 1,
+    };
+    let dpad_up_button = Toggle {
+        title: "Dpad Up Button",
+        value: 0,
+        max: 1,
+    };
+    let dpad_down_button = Toggle {
+        title: "Dpad Down Button",
+        value: 0,
+        max: 1,
+    };
+    let dpad_left_button = Toggle {
+        title: "Dpad Left Button",
+        value: 0,
+        max: 1,
+    };
+    let dpad_right_button = Toggle {
+        title: "Dpad Right Button",
+        value: 0,
+        max: 1,
+    };
+    vec![
+        a_button,
+        b_button,
+        x_button,
+        y_button,
+        l_button,
+        r_button,
+        zl_button,
+        zr_button,
+        dpad_up_button,
+        dpad_down_button,
+        dpad_left_button,
+        dpad_right_button,
+    ]
+}
+
+fn new_on_off<'a>() -> Vec<Toggle<'a>> {
+    let true_toggle = Toggle {
+        title: "True",
+        value: 0,
+        max: 1,
+    };
+    let false_toggle = Toggle {
+        title: "False",
+        value: 0,
+        max: 1,
+    };
+    vec![true_toggle, false_toggle]
+}
+
+fn new_toggle_table<'a>(items: Vec<Toggle<'a>>) -> StatefulTable<Toggle<'a>> {
+    StatefulTable::with_items(NX_SUBMENU_ROWS, NX_SUBMENU_COLUMNS, items)
 }
