+++ training_mod_tui/src/structures/mod.rs
