+++ training_mod_tui/src/structures/stateful_slider.rs
