+++ training_mod_tui/tests/test_submenu.rs
