+++ training_mod_tui/tests/test_toggle.rs
