--- training_mod_tui_2/Cargo.toml
@@ -1,18 +1,0 @@
-[package]
-name = "training_mod_tui_2"
-version = "0.1.0"
-edition = "2021"
-
-[lib]
-doctest = false
-
-[dependencies]
-itertools = "0.11.0"
-ratatui = { git = "https://github.com/tonogdlp/ratatui.git", branch = "single-cell", default-features = false}
-serde = { version = "1.0.188", features = ["derive"] }
-serde_json = "1.0.106"
-crossterm = { version = "0.22.1", optional = true }
-
-[features]
-default = []
-has_terminal = ["crossterm", "ratatui/crossterm"]
