--- training_mod_tui_2/src/lib.rs
@@ -1,10 +1,0 @@
-mod containers;
-pub use containers::*;
-mod structures;
-pub use structures::*;
-
-pub mod ui;
-pub use crate::ui::*;
-
-pub const NX_SUBMENU_ROWS: usize = 8;
-pub const NX_SUBMENU_COLUMNS: usize = 4;
